configuring of license wurde ausgelagert da es an mehreren stellen gebraucht wird
