+++ src/views/TabDataProtection.vue
@@ -94,7 +94,7 @@
           </h2>
           <p>
             Durch die obigen Angaben werden deine Daten unter der folgenden Lizenz gespeichert:
-            <strong>{{licensePTR}}</strong>
+            <strong>{{exportedLicensePTR}}</strong>
           </p>
         </ion-label>
         </div>
@@ -134,6 +134,11 @@ import {
   setLicense, setSecondLanguage, setZipCode
 } from "@/scripts/UserSettingsStorage";
 import 'firebase/firestore';
+import {evaluateLicenseAndDeactivations, evaluateButtonSettingsFromLicense,
+        saveLocalData, uploadLicense,exportedLicensePTR, downloadLicense,
+        loadLocalData, options, isComerciallyUseAllowed, isComerciallyUseAllowedDeactivated,
+        isMentioningActivated, isMentioningActivatedDeactivated, isRemixingAllowedDeactivated,
+        isRemixingAllowed, isSharingAllowed, isSharingAllowedDeactivated, optionChanged} from "@/scripts/LicenseSettings";
 export default defineComponent({
 
 
@@ -146,195 +151,13 @@ export default defineComponent({
     // multi-lingual support
     const { t } = useI18n();
 
-//
-// TODO:
-// load these values from the server! and change the toggles
-//
-
-//TODO upload data
-
-  const licensePTR = ref('CC0 1.0');
-
-  const options = new Map<string, boolean>([
-    ["isMentioningActivated", false],
-    ["isComerciallyUseAllowed", true],
-    ["isRemixingAllowed", true],
-    ["isSharingAllowed", true],
-  ]);
-
-//since the checked property is one way bounded, these are only used for initialisation
-  const isMentioningActivated = ref(false);
-  const isComerciallyUseAllowed = ref(true);
-  const isRemixingAllowed = ref(true);
-  const isSharingAllowed = ref(true);
-
-  const isMentioningActivatedDeactivated = ref(false);
-  const isComerciallyUseAllowedDeactivated = ref(true);
-  const isRemixingAllowedDeactivated = ref(true);
-  const isSharingAllowedDeactivated = ref(true);
-
-    //7 different licenses
-    const evaluateLicenseAndDeactivations = ()=>{
-      if(options.get("isMentioningActivated")){
-        //user does not want attribution
-        //deactivate all other toggles
-        isComerciallyUseAllowedDeactivated.value = false;
-        isRemixingAllowedDeactivated.value = false;
-        isSharingAllowedDeactivated.value = false;
-        //set license
-        licensePTR.value = "CC BY 4.0";
-
-        //if one of the lower toggles is unchecked, deactivate the top one
-        if(!(options.get("isComerciallyUseAllowed") && options.get("isRemixingAllowed") && options.get("isSharingAllowed"))){
-          isMentioningActivatedDeactivated.value = true;
-
-          licensePTR.value = "something else";
-          //combinations of the last 3 values
-          if(options.get("isComerciallyUseAllowed") && options.get("isRemixingAllowed") && !options.get("isSharingAllowed")){
-            //BY-SA
-            isRemixingAllowedDeactivated.value = true;
-            licensePTR.value = "CC BY-SA 4.0";
-          }else if(options.get("isComerciallyUseAllowed") && !options.get("isRemixingAllowed") && options.get("isSharingAllowed")){
-            //BY-ND
-            isSharingAllowedDeactivated.value = true;
-            licensePTR.value = "CC BY-ND 4.0";
-          }else if(!options.get("isComerciallyUseAllowed") && options.get("isRemixingAllowed") && options.get("isSharingAllowed")){
-            //BY-NC
-            licensePTR.value = "CC BY-NC 4.0";
-          }else if(!options.get("isComerciallyUseAllowed") && options.get("isRemixingAllowed") && !options.get("isSharingAllowed")){
-            //BY-NC-SA
-            isRemixingAllowedDeactivated.value =true;
-            licensePTR.value = "CC BY-NC-SA 4.0";
-          }else if(!options.get("isComerciallyUseAllowed") && !options.get("isRemixingAllowed") && options.get("isSharingAllowed")){
-            //BY-NC-ND
-            isSharingAllowedDeactivated.value = true;
-            licensePTR.value = "CC BY-NC-ND 4.0";
-          }else{
-            console.log("ERROR");
-          }
-        }else{
-          isMentioningActivatedDeactivated.value = false;
-        }
-
-      }else{
-        isComerciallyUseAllowedDeactivated.value = true;
-        isRemixingAllowedDeactivated.value = true;
-        isSharingAllowedDeactivated.value = true;
-        licensePTR.value = "CC0 1.0";
-      }
-    }
-
-
-    const evaluateButtonSettingsFromLicense = ()=>{
-      if(licensePTR.value == "CC0 1.0"){
-        return;//leave everything default
-      }
-      options.set("isMentioningActivated", true);
-      isMentioningActivated.value = true;
-      if(licensePTR.value.includes("NC")){
-        options.set("isComerciallyUseAllowed", false);
-        isComerciallyUseAllowed.value = false;
-      }
-      if(licensePTR.value.includes("SA")){
-        options.set("isSharingAllowed", false);
-        isSharingAllowed.value = false;
-      }
-      if(licensePTR.value.includes("ND")){
-        options.set("isRemixingAllowed", false);
-        isRemixingAllowed.value = false;
-      }
-      evaluateLicenseAndDeactivations();
-    }
-
-    const uploadLicense = ()=>{
-
-    const db = firebase.firestore();
-      const currentUser = firebase.auth().currentUser;
-      if (currentUser == null) return;
-      const userUID = currentUser.uid;
-
-      console.log("Start uploading");
-
-      db.collection("users").doc(userUID).set({
-        license: licensePTR.value
-      },{merge: true});
-      console.log("wrote to database");
-    }
-
-
-    const downloadLicense = () =>{ //nicht mehr benötigt eigentlich im moment, kann aber vielleicht noch sinnvoll werden
-    //wird sehr wohl noch verwendet. wenn man lokal keine daten hat, muss man die von der datenbank nehmen
-    const db = firebase.firestore();
-      const currentUser = firebase.auth().currentUser;
-      if (currentUser == null) return;
-      const userUID = currentUser.uid;
-
-/*  const options = new Map<string, boolean>([
-    ["isMentioningActivated", false],
-    ["isComerciallyUseAllowed", true],
-    ["isRemixingAllowed", true],
-    ["isSharingAllowed", true],
-  ]);
-
-  */
-      db.collection("users").doc(userUID).get().then((doc)=>{
-        console.log(doc.get('license'));
-        const license = doc.get('license');
-        if(license == "CC0 1.0"){
-          return;//leave everything default
-        }
-        options.set("isMentioningActivated", true);
-        isMentioningActivated.value = true;
-        if(license.includes("NC")){
-          options.set("isComerciallyUseAllowed", false);
-          isComerciallyUseAllowed.value = false;
-        }
-        if(license.includes("SA")){
-          options.set("isSharingAllowed", false);
-          isSharingAllowed.value = false;
-        }
-        if(license.includes("ND")){
-          options.set("isRemixingAllowed", false);
-          isRemixingAllowed.value = false;
-        }
-        evaluateLicenseAndDeactivations();
-      });
-    }
-
-    const loadLocalData = async () =>{
-      const user = firebase.auth().currentUser;
-      if (user == null) return;
-      await loadUserSettings();
-      licensePTR.value=getLicense();
-      evaluateButtonSettingsFromLicense();
-    }
-
-    const saveLocalData = async () =>{
-      const user = firebase.auth().currentUser;
-      if (user == null) return;
-      setLicense(licensePTR.value);
-      const toast = await toastController
-          .create({
-            message: 'Your license Changes have been saved',
-            duration: 1500
-          })
-      return toast.present();
-    }
-
-    const optionChanged = (event: any)=>{
-      options.set(event.target.name, event.target.checked);
-      evaluateLicenseAndDeactivations();
-      setLicense(licensePTR.value);
-      saveLocalData();
-      uploadLicense();
-    }
 
     loadLocalData();
     uploadLicense();//überschreibt datenbank
 
 
 
-    return { t, licensePTR, optionChanged, 
+    return { t, optionChanged,
     isSharingAllowedDeactivated,
     isRemixingAllowedDeactivated,
     isComerciallyUseAllowedDeactivated,
@@ -342,7 +165,7 @@ export default defineComponent({
     isMentioningActivated,
     isComerciallyUseAllowed,
     isRemixingAllowed,
-    isSharingAllowed
+    isSharingAllowed, exportedLicensePTR,
     }
   }
 })
